requirements.txt
fastapi
uvicorn[standard]
python-multipart
faiss-cpu
sentence-transformers
pypdf
openai
numpy
python-dotenv
pytest
pytest-asyncio
httpx
